Add a delay when sending to the device.
Since the receive buffer on the target devince might have a limited size (e.g. 64 byte for Arduino Seral), data is send in 8 byte chunks with a delay of 100ms inbetween.
Also added a custom delay function that is called by ubrb. This can be Arduino's delay() or something special for devices that need to tick the USB/serial subsystem (like Digispark's SerialUSB). This replaces the busy waiting.